backend/requirements.txt
pypdf
django
djangorestframework
django-cors-headers
pandas
joblib
scikit-learn